+++ src/core_modules/capture-core/components/DataEntries/SingleEventRegistrationEntry/SingleEventRegistrationEntry.types.js
@@ -16,4 +16,8 @@ export type ContainerProps = {|
     id: string,
 |};
 
+export type DispatchProps = {|
+    onCancel: () => void,
+|};
+
 export type MapStateToProps = (ReduxState, ContainerProps) => StateProps;