+++ src/core_modules/capture-core/components/Pages/New/RegistrationDataEntry/RegistrationDataEntry.component.js
@@ -1,6 +1,7 @@
 // @flow
 import React, { type ComponentType, useContext, useCallback } from 'react';
 import { useDispatch } from 'react-redux';
+import { useHistory } from 'react-router';
 import i18n from '@dhis2/d2-i18n';
 import { Button, colors, spacers } from '@dhis2/ui';
 import { Grid, withStyles } from '@material-ui/core';
@@ -13,7 +14,7 @@ import { TrackedEntityTypeSelector } from '../../../TrackedEntityTypeSelector';
 import { DataEntryWidgetOutput } from '../../../DataEntryWidgetOutput/DataEntryWidgetOutput.container';
 import { ResultsPageSizeContext } from '../../shared-contexts';
 import { navigateToEnrollmentOverview } from '../../../../actions/navigateToEnrollmentOverview/navigateToEnrollmentOverview.actions';
-import { useLocationQuery } from '../../../../utils/routing';
+import { buildUrlQueryString, useLocationQuery } from '../../../../utils/routing';
 import { EnrollmentRegistrationEntryWrapper } from '../EnrollmentRegistrationEntryWrapper.component';
 import { useCurrentOrgUnitId } from '../../../../hooks/useCurrentOrgUnitId';
 
@@ -94,18 +95,29 @@ const RegistrationDataEntryPlain = ({
     teiId,
     trackedEntityInstanceAttributes,
 }: Props) => {
+    const { push } = useHistory();
     const { resultsPageSize } = useContext(ResultsPageSizeContext);
     const { scopeType, programName, trackedEntityName } = useScopeInfo(selectedScopeId);
     const titleText = useScopeTitleText(selectedScopeId);
     const currentOrgUnitId = useCurrentOrgUnitId();
 
+    const onCancel = useCallback(() => {
+        let url;
+        if (scopeType === scopeTypes.TRACKER_PROGRAM) {
+            url = buildUrlQueryString({ programId: selectedScopeId, orgUnitId: currentOrgUnitId });
+        } else {
+            url = buildUrlQueryString({ orgUnitId: currentOrgUnitId });
+        }
+        return push(`/?${url}`);
+    }, [currentOrgUnitId, push, scopeType, selectedScopeId]);
+
     const handleRegistrationScopeSelection = (id) => {
         setScopeId(id);
     };
 
-    const renderDuplicatesDialogActions = useCallback((onCancel, onSave) => (
+    const renderDuplicatesDialogActions = useCallback((callbackOnCancel, onSave) => (
         <DialogButtons
-            onCancel={onCancel}
+            onCancel={callbackOnCancel}
             onSave={onSave}
         />
     ), []);
@@ -180,6 +192,7 @@ const RegistrationDataEntryPlain = ({
                                     teiId={teiId}
                                     selectedScopeId={selectedScopeId}
                                     onSave={onSaveWithEnrollment}
+                                    onCancel={onCancel}
                                     saveButtonText={(trackedEntityTypeNameLC: string) => i18n.t('Save {{trackedEntityTypeName}}', {
                                         trackedEntityTypeName: trackedEntityTypeNameLC,
                                         interpolation: { escapeValue: false },
@@ -232,6 +245,7 @@ const RegistrationDataEntryPlain = ({
                                     id={dataEntryId}
                                     selectedScopeId={selectedScopeId}
                                     orgUnitId={currentOrgUnitId}
+                                    onCancel={onCancel}
                                     saveButtonText={i18n.t('Save {{trackedEntityName}}', {
                                         trackedEntityName,
                                         interpolation: { escapeValue: false },