+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/StageDetail/StageDetail.component.js
@@ -16,7 +16,7 @@ import { colors,
     IconAdd16,
     Tooltip,
 } from '@dhis2/ui';
-import { ConditionalTooltip } from 'capture-core/components/ConditionalTooltip';
+import { ConditionalTooltip } from 'capture-core/components/Tooltips/ConditionalTooltip';
 import { sortDataFromEvent } from './hooks/sortFuntions';
 import { useComputeDataFromEvent, useComputeHeaderColumn, formatRowForView } from './hooks/useEventList';
 import { DEFAULT_NUMBER_OF_ROW, SORT_DIRECTION } from './hooks/constants';
@@ -66,6 +66,7 @@ const StageDetailPlain = (props: Props) => {
         dataElements,
         hideDueDate = false,
         repeatable = false,
+        enableUserAssignment = false,
         onEventClick,
         onViewAll,
         onCreateNew,
@@ -76,7 +77,7 @@ const StageDetailPlain = (props: Props) => {
         sortDirection: SORT_DIRECTION.DESC,
     };
     const { stage } = getProgramAndStageForProgram(programId, stageId);
-    const headerColumns = useComputeHeaderColumn(dataElements, hideDueDate, stage?.stageForm);
+    const headerColumns = useComputeHeaderColumn(dataElements, hideDueDate, enableUserAssignment, stage?.stageForm);
     const { loading, value: dataSource, error } = useComputeDataFromEvent(dataElements, events);