+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsBase/TemplateMaintenance/ExistingTemplateDialog.component.js
@@ -1,6 +1,6 @@
 // @flow
 import * as React from 'react';
-import { Dialog } from '@material-ui/core';
+import { Modal } from '@dhis2/ui';
 import { ExistingTemplateContents } from './ExistingTemplateContents.component';
 
 type Props = {
@@ -16,16 +16,21 @@ export const ExistingTemplateDialog = (props: Props) => {
         onSaveTemplate,
     } = props;
 
+    if (!open) {
+        return null;
+    }
+
     return (
-        <Dialog
-            open={open}
+        <Modal
+            hide={!open}
             onClose={onClose}
-            fullWidth
+            position={'center'}
+            dataTest={'existing-template-dialog'}
         >
             <ExistingTemplateContents
                 onSaveTemplate={onSaveTemplate}
                 onClose={onClose}
             />
-        </Dialog>
+        </Modal>
     );
 };