+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsBase/TemplateSelectorChip.component.js
@@ -3,6 +3,7 @@ import * as React from 'react';
 import { Chip } from '@dhis2/ui';
 import { TemplateSelectorChipContent } from './TemplateSelectorChipContent.component';
 import type { WorkingListTemplate } from './workingListsBase.types';
+import { TooltipForChip } from '../../Tooltips/TooltipForChip';
 
 type PassOnProps = {
     currentListIsModified: boolean,
@@ -13,10 +14,17 @@ type Props = {
     template: WorkingListTemplate,
     currentTemplateId: string,
     onSelectTemplate: Function,
+    maxCharacters?: number,
 };
 
 export const TemplateSelectorChip = (props: Props) => {
-    const { template, currentTemplateId, onSelectTemplate, ...passOnProps } = props;
+    const {
+        template,
+        currentTemplateId,
+        onSelectTemplate,
+        maxCharacters = 30,
+        ...passOnProps
+    } = props;
     const { name, id } = template;
 
     const selectTemplateHandler = React.useCallback(() => {
@@ -27,16 +35,29 @@ export const TemplateSelectorChip = (props: Props) => {
     ]);
 
     return (
-        <Chip
-            dataTest="workinglist-template-selector-chip"
-            selected={id === currentTemplateId}
+        <TooltipForChip
+            content={name}
+            placement={'top'}
+            openDelay={800}
+            enabled={name.length > maxCharacters}
             onClick={selectTemplateHandler}
         >
-            <TemplateSelectorChipContent
-                {...passOnProps}
-                text={name}
-                isSelectedTemplate={id === currentTemplateId}
-            />
-        </Chip>
+            <Chip
+                marginTop={0}
+                marginBottom={0}
+                marginLeft={0}
+                marginRight={0}
+                dataTest="workinglist-template-selector-chip"
+                selected={id === currentTemplateId}
+            >
+                <TemplateSelectorChipContent
+                    {...passOnProps}
+                    text={name}
+                    maxCharacters={maxCharacters}
+                    isSelectedTemplate={id === currentTemplateId}
+                />
+            </Chip>
+        </TooltipForChip>
     );
 };
+