+++ src/core_modules/capture-core/utils/reactQueryHelpers/reactQueryHelpers.const.js
@@ -1,4 +1,4 @@
-
 // @flow
 
 export const ReactQueryAppNamespace = 'capture';
+export const IndexedDBNamespace = 'indexedDB';